% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestparcoords.R
\name{bestparcoord}
\alias{bestparcoord}
\title{Create best parallel coords}
\usage{
bestparcoord(data)
}
\arguments{
\item{data}{A dataframe with appropriately dealt with columns}
}
\value{
The column names in the best order
}
\description{
Returns the order to put columns in for a starting point for the best parallel coordinates.
Also creates a preliminary plot with the returned order.
Assumes columns are already scaled appropriately and are not categorical.
}
